Reorganizes DB methods into db/methods and cleans up db/index.js
