+++ src/db/methods/index.js
@@ -1,4 +1,10 @@
 module.exports = {
   generateFakeData: require('./generateFakeData'),
-  topBuys: require('./topBuys.js'),
+  topBuys: require('./topBuys'),
+  topSells: require('./topSells'),
+  openPosition: require('./openPosition'),
+  updatePosition: require('./updatePosition'),
+  resolvePosition: require('./resolvePosition'),
+  closeOrder: require('./closeOrder'),
+  processOrder: require('./processOrder'),
 };
