+++ src/db/methods/openPosition.js
@@ -1,0 +1,25 @@
+const { Position } = require('..');
+const { elasticClient } = require('../../../workers/elasticSetup');
+
+// Open a new user position
+module.exports = openPosition = ({userId, price, volume, type}) => {
+  type = (type === 'BUY') ? 'long' : 'short';
+  // create a position w/ obj passed in
+  // save to DB
+  elasticClient.index({
+    type,
+    id: userId,
+    index: 'positions',
+    body: {
+      price,
+      volume
+    }
+  });
+  return Position.create({
+    userId,
+    price,
+    volume,
+    type,
+    orders: [{ price, volume }],
+  });
+};
