+++ src/db/methods/processOrder.js
@@ -1,0 +1,91 @@
+const { Buy, Sell, Position } = require('..');
+const { topBuys, topSells, resolvePosition, processOrder, closeOrder } = require('.');
+
+// Handle an incoming order from the queue/HTTP request
+module.exports = ({ type, order }) => {
+  let { userId, volume, price } = order;
+  // console.log('SAW', volume, price);
+  // if the order is to buy, try to find a matching sell order or add it to the buy list
+  if (type === 'BUY') {
+    // console.log('HI BUY');
+    // load the next 10 sell orders in line
+    topSells(top => {
+      // console.log('FOUND: ', top[0]);
+      // if price is less than the best sell price listed, we know there's no match
+      if (price < top[0].price) {
+        // console.log('below best price, creating new order');
+        Buy.create(order).then(result => {
+          // console.log(result)
+        });
+      // if price is greater than the best sell price listed, we match the order
+      } else {
+        // console.log('above best price, resolving order');
+        let remainingVol = volume;
+        let i = 0;
+        let reprocess = false;
+        // we close listings until the entire order volume has been fulfilled
+        while (remainingVol > 0 && i < top.length) {
+          // it's possible that the next price in the list is higher than the previous, so we check
+          if (price >= top[i].price) {
+            // console.log('remaining: ', remainingVol);
+            remainingVol = closeOrder(top[i], remainingVol, 'SELL');
+            // console.log('new volume is: ', remainingVol);
+            // console.log('Im gonna write: ', volume - remainingVol);
+            resolvePosition({ userId, price: top[i].price, volume: volume - remainingVol }, type);
+            i++;
+          // if the order price no longer beats the listed price, we set the reprocess flag to true
+          } else {
+            reprocess = true;
+            break;
+          }
+        }
+        // if the top 10 listings did not fulfill the order, repeat the process
+        if (reprocess || remainingVol) {
+          order.volume = remainingVol;
+          processOrder({ type, order });
+        }
+      }
+    });
+  // if the order is to sell, try to find a matching buy order or add it to the sell list
+  } else if (type === 'SELL') {
+    // console.log('HI SELL');
+    // load the next 10 buy orders in line
+    topBuys(top => {
+      // console.log('FOUND: ', top[0]);
+      // if price is greater than the best buy price listed, we know there's no match
+      if (price > top[0].price) {
+        // console.log('above best price, creating new order');        
+        Sell.create(order).then(result => {
+          // console.log('SELL RESULT: ', result)
+        });
+      // if price is less than the best buy price listed, we match the order
+      } else {
+        // console.log('beats best price, resolving order');        
+        let remainingVol = volume;
+        let i = 0;
+        let reprocess = false;
+        // we close listings until the entire order volume has been fulfilled
+        while (remainingVol > 0 && i < top.length) {
+          // it's possible that the next price in the list is lower than the previous, so we check
+          if (price <= top[i].price) {
+            // console.log('remaining: ', remainingVol);
+            remainingVol = closeOrder(top[i], remainingVol, 'BUY');
+            // console.log('new volume is: ', remainingVol);
+            // console.log('Im gonna write: ', volume - remainingVol);
+            resolvePosition({ userId, price: top[i].price, volume: volume - remainingVol }, type);
+            i++;
+          // if the order price no longer beats the listed price, we set the reprocess flag to true
+          } else {
+            reprocess = true;
+            break;
+          }
+        }
+        // if the top 10 listings did not fulfill the order, repeat the process
+        if (reprocess || remainingVol) {
+          order.volume = remainingVol;
+          processOrder({ type, order });
+        }
+      }
+    });
+  }
+};
