+++ src/db/methods/topBuys.js
@@ -1,7 +1,8 @@
 const { Buy, sequelize } = require('..');
 
+// Find the first 10 open orders in the Buy table
 module.exports = (callback) => {
-  console.log('starting sort');
+  // console.log('starting sort');
   return Buy
     .max('price')
     .then(max => Buy.findAll({
