+++ src/db/methods/topSells.js
@@ -1,0 +1,14 @@
+const { Sell, sequelize } = require('..');
+
+// Find the first 10 open orders in the Sell table
+module.exports = callback => {
+  // console.log('starting sort');
+  return Sell
+    .min('price')
+    .then(min => Sell.findAll({
+      limit: 10,
+      where: { price: min }, 
+      order: [[sequelize.col('price'), 'ASC'], [sequelize.col('createdAt'), 'ASC']]
+    }))
+    .then(results => callback(results));
+};
